HualuCup2021/requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
albumentations==1.0.3
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==21.2.0
blis==0.7.4
cached-property==1.5.2
cachetools==4.2.2
catalogue==2.0.4
certifi==2018.1.18
chardet==3.0.4
charset-normalizer==2.0.3
click==7.1.2
contextvars==2.4
cryptography==2.1.4
cupy-cuda110==9.0.0b3
cycler==0.10.0
cymem==2.0.5
Cython==0.29.24
dataclasses==0.8
decorator==4.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
fastrlock==0.6
filelock==3.0.12
fsspec==2021.7.0
future==0.18.2
google-auth==1.33.0
google-auth-oauthlib==0.4.4
grpcio==1.38.1
h5py==3.1.0
huggingface-hub==0.0.12
idna==2.6
idna-ssl==1.1.0
imageio==2.9.0
immutables==0.15
importlib-metadata==4.6.1
Jinja2==3.0.1
joblib==1.0.1
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
lmdb==1.2.1
lmdbdict==0.2.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
multidict==5.1.0
murmurhash==1.0.5
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.5.3.56
opencv-python-headless==4.5.3.56
packaging==21.0
pandas==1.1.5
pathy==0.6.0
Pillow==8.3.1
preshed==3.0.5
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
pydantic==1.8.2
pygobject==3.26.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytorch-lightning==1.2.6
pytz==2021.1
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.3.1
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
scikit-image==0.17.2
scipy==1.5.4
SecretStorage==2.3.1
six==1.11.0
smart-open==5.1.0
spacy==3.1.0
spacy-legacy==3.0.8
spacy-pkuseg==0.0.28
srsly==2.4.1
ssh-import-id==5.7
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thinc==8.0.7
tifffile==2020.9.3
timm==0.4.12
tokenizers==0.10.3
torch==1.7.1+cu110
torchaudio==0.7.2
torchmetrics==0.4.1
torchvision==0.8.2+cu110
tqdm==4.61.2
transformers==4.8.2
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.22
wasabi==0.8.2
Werkzeug==2.0.1
yacs==0.1.8
yarl==1.6.3
zh-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.1.0/zh_core_web_sm-3.1.0-py3-none-any.whl
zipp==3.5.0
